#89 Moved spline_matrix() to spline.h
