+++ graphics/spline_curve.h
@@ -1,17 +1,6 @@
 #pragma once
 
-#include "transforms.h"
-
-// Coefficient matrix used to calculate a Spline curve
-inline TMatrix spline_matrix()
-{
-    return TMatrix(
-        { -1.0/6.0,      0.5,    -0.5, 1.0/6.0 },
-        {      0.5,     -1.0,     0.5,     0.0 },
-        {     -0.5,      0.0,     0.5,     0.0 },
-        {  1.0/6.0,  4.0/6.0, 1.0/6.0,     0.0 }
-    );
-}
+#include "spline.h"
 
 // Matrix used to calculate initial deltas of forward differences
 inline TMatrix delta_matrix()
